:recycle:  PasswordEncoder 추가로 회원가입, 로그인, 회원정보 수정, 회원정보 삭제 기능 구조 변경
회원가입은 비밀번호를 PasswordEncoder로 암호화하여서 DB에 저장되게 수정했다.

로그인은 email로 찾은 회원 정보의 비밀번호와 요청 데이터의 비밀번호 사이의 동일 여부를 검증하도록 수정했다.

회원정보 수정은 현재 비밀번호와 새 비밀번호를 요청데이터로 받을 수 있고 현재 비밀 번호 검증을 거친 후 새 비밀 번호로 수정할 수 있다.

회원정보 삭제는 비밀번호를 요청데이터로 반드시 받아 회원 검증을 거친 후 회원 정보를 삭제할 수 있다.

+++ src/main/java/com/example/schedulerjpa/service/UserServiceImpl.java
@@ -1,5 +1,6 @@
 package com.example.schedulerjpa.service;
 
+import com.example.schedulerjpa.config.PasswordEncoder;
 import com.example.schedulerjpa.dto.*;
 import com.example.schedulerjpa.entity.User;
 import com.example.schedulerjpa.repository.UserRepository;
@@ -18,9 +19,15 @@ public class UserServiceImpl implements UserService{
     private final UserRepository userRepository;
 
     @Override
-    public SignUpUserResponseDto signUpUser(String userName, String email, String password) {
+    public SignUpUserResponseDto signUpUser(SignUpUserRequestDto signUpUserRequestDto) {
 
-        User user = new User(userName, email, password);
+        PasswordEncoder passwordEncoder = new PasswordEncoder();
+
+        String encodedPassword = passwordEncoder.encode(signUpUserRequestDto.getPassword());
+        User user = new User(
+                signUpUserRequestDto.getUserName(),
+                signUpUserRequestDto.getEmail(),
+                encodedPassword);
 
         User savedUser = userRepository.save(user);
 
@@ -33,9 +40,14 @@ public class UserServiceImpl implements UserService{
     }
 
     @Override
-    public LoginResponseDto login(String email, String password) {
-        User user = userRepository.findUserByEmailAndPasswordOrElseThrow(email, password);
-        return new LoginResponseDto(user.getId(), user.getUserName());
+    public LoginResponseDto login(LoginRequestDto loginRequestDto) {
+        PasswordEncoder passwordEncoder = new PasswordEncoder();
+        User user = userRepository.findUserByUserEmailOrElseThrow(loginRequestDto.getEmail());
+
+     if(passwordEncoder.matches(loginRequestDto.getPassword(), user.getPassword())) {
+         return new LoginResponseDto(user.getId(), user.getUserName());
+     } else {throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+     }
     }
 
     @Override
@@ -58,6 +70,15 @@ public class UserServiceImpl implements UserService{
     public UpdateUserResponseDto updateUser(Long id, UpdateUserRequestDto requestDto) {
         User user = userRepository.findUserByIdOrElseThrow(id);
 
+        if (requestDto.getPresentPassword() != null && !requestDto.getPresentPassword().isBlank()) {
+            PasswordEncoder passwordEncoder = new PasswordEncoder();
+           if (passwordEncoder.matches(requestDto.getPresentPassword(), user.getPassword())) {
+               if(requestDto.getNewPassword() != null && !requestDto.getNewPassword().isBlank()) {
+                   user.setPassword(requestDto.getNewPassword());
+               }
+           } else {throw new ResponseStatusException(HttpStatus.BAD_REQUEST);}
+        }
+
         if (requestDto.getUserName() != null && !requestDto.getUserName().isBlank()) {
             user.setUserName(requestDto.getUserName());
         }
@@ -70,9 +91,11 @@ public class UserServiceImpl implements UserService{
     }
 
     @Override
-    public void deleteUserById(Long id) {
+    public void deleteUserById(Long id, String password) {
         User user = userRepository.findUserByIdOrElseThrow(id);
-
-        userRepository.delete(user);
+        PasswordEncoder passwordEncoder = new PasswordEncoder();
+        if(passwordEncoder.matches(password, user.getPassword())) {
+            userRepository.delete(user);
+        } else {throw new ResponseStatusException(HttpStatus.BAD_REQUEST);}
     }
 }
